pyproject.toml
[tool.poetry]
name = "mrhelper"
version = "0.0.1"
description = "a simple movie-robot helper"
authors = ['yanyao']
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.0"
nonebot_plugin_apscheduler = "^0.2.0"
nonebot_plugin_gocqhttp = "^0.6.2"
nonebot-adapter-onebot = "^2.1.4"
nb-cli = "^0.6.7"
nonebot2 = "^2.0.0rc1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.7"

[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_gocqhttp"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
